_legacy_archive/requirements.txt
Django>=5.0
google-generativeai
python-dotenv
requests
